Display files in a table
This allows for more information to be displayed. The table columns are
filename, size, and modified date.

This commit adds a new optional parameter to /manga/information route.
This parameter's value is restricted to ascending and descending.

As the previous paragraph suggests, the table can be sorted in ascending
and descending order based on the filename.