Allow library creation without waiting for request.
Prior to this commit, the library would be scanned using
the request thread and would result in noticeable wait times.
There was also the problem that the request could timeout.

I have removed the scan upon creation and also created a job
that will be run in a separate terminal with 'php artisan queue:work'
This ensures that the web ui will remain responsive.

Additionally, the web ui now has a modal that will show progress
to the user and supports updating and deleting several libraries
in one request.

There are also small bug fixes in other files.