Update public directory. Read full commit for nginx config changes.
In your nginx config, you will need to update the root path from
/var/www/mangapie to /var/www/mangapie/public.

At some point, index.php was moved from public to the root of the project.
To avoid headachces on Windows and Apache, I have decided to move it back to public.

public/.htaccess has also been updated to reflect the 5.8 branch of laravel.