Fix autofill errors.
MangaUpdates::autofillFromId was incorrectly calling forceDelete on the
wrong objects. Additionally, the genre logic would have failed in the
future if the genre did not exist as there is no default for a missing
description. It is more verbose to let it fail when accessing the
properties of a non-object (null).

MangaEditController@putAutoFill was calling Manga::firstOrFail which is
not needed for a table where every item is unique.