Add mail option for notifications.
There is an option to enable or disable mailing notifications through
the .env file. Be sure to run `php artisan config:cache` and restart any
running queue workers and watchers.

The ReaderController@index method now handles a query parameter named
`notification` that will dismiss the DatabaseNotification with the
matching id.